Moved several types to @memberjunction/ai-core-plus to make it possible to use these types on client + server --- anywhere as reference types. Previously some were in ai-prompts and some in aiengine which are server-only
