+++ metadata/prompts/.prompts.json
@@ -121,7 +121,7 @@
       "ID": "B4053E4F-165C-4AED-8E14-3A0A25A28EFE"
     },
     "sync": {
-      "lastModified": "2025-06-25T21:23:35.154Z",
+      "lastModified": "2025-06-25T21:41:36.461Z",
       "checksum": "cab30b4b034ad25dc3277a069011d03437bc361fa69af1dff198516b14d800d5"
     }
   },
