+++ packages/AI/Core/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@memberjunction/ai",
   "version": "2.57.0",
-  "description": "MemberJunction: AI Module - LLMs and other AI services wrapped and abstracted for use in MemberJunction",
+  "description": "MemberJunction: AI - core components for abstracting LLMs and other AI model types that are usable anywhere without ANY other MJ dependencies past @memberjunction/global which itself has zero additional dependencies.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -16,7 +16,6 @@
   "license": "ISC",
   "dependencies": {
     "@memberjunction/global": "2.57.0",
-    "@memberjunction/core-entities": "2.57.0",
     "dotenv": "^16.4.1",
     "rxjs": "^7.8.1"
   },
