+++ packages/AI/Core/src/index.ts
@@ -9,4 +9,3 @@ export * from './generic/embed.types';
 export * from './generic/baseEmbeddings'; 
 export * from './generic/baseAudio';
 export * from './generic/baseVideo';
-export * from './generic/prompt.types';
