+++ packages/AI/CorePlus/tsconfig.json
@@ -1,0 +1,26 @@
+{
+    "compilerOptions": {
+      "module": "commonjs",
+      "target": "es2020",
+      "declaration": true,
+      "declarationMap": true,
+      "outDir": "dist",
+      "rootDir": "src",
+      "sourceMap": true,
+      "experimentalDecorators": true,
+      "resolveJsonModule": true,
+      "esModuleInterop": true,
+      "lib": ["es2020", "esnext.asynciterable"],
+      "moduleResolution": "node",
+      "removeComments": true,
+      "noImplicitThis": true,
+      "noImplicitReturns": true,
+      "noFallthroughCasesInSwitch": true,
+      "allowSyntheticDefaultImports": true,
+      "emitDecoratorMetadata": true,
+      "skipLibCheck": true
+    },
+    "exclude": ["node_modules"],
+    "include": ["./src/**/*.ts"]
+  }
+
