+++ packages/AI/Engine/src/index.ts
@@ -1,2 +1,1 @@
-export * from './AIEngine';
-export * from './agent-types';
+export * from './AIEngine';
