+++ packages/AI/Prompts/package.json
@@ -19,6 +19,7 @@
     "@memberjunction/global": "2.57.0",
     "@memberjunction/core-entities": "2.57.0",
     "@memberjunction/ai": "2.57.0",
+    "@memberjunction/ai-core-plus": "2.57.0",
     "@memberjunction/aiengine": "2.57.0",
     "@memberjunction/templates": "2.57.0",
     "@memberjunction/templates-base-types": "2.57.0",
