+++ packages/AI/Prompts/src/index.ts
@@ -1,7 +1,5 @@
 import { LoadAIPromptRunner } from './AIPromptRunner';
 LoadAIPromptRunner(); // Ensure AIPromptRunner class isn't tree-shaken
 
-export * from './AIPromptCategoryExtended';
 export * from './AIPromptRunner';
 export * from './SystemPlaceholders';
-export * from './types';
