+++ packages/Actions/CoreActions/package.json
@@ -22,6 +22,7 @@
     "@memberjunction/global": "2.57.0",
     "@memberjunction/core": "2.57.0",
     "@memberjunction/core-entities": "2.57.0",
+    "@memberjunction/ai-core-plus": "2.57.0",
     "@memberjunction/actions": "2.57.0",
     "@memberjunction/communication-engine": "2.57.0",
     "@memberjunction/external-change-detection": "2.57.0",
