+++ packages/Actions/Engine/package.json
@@ -24,6 +24,7 @@
     "@memberjunction/actions-base": "2.57.0",
     "@memberjunction/core-entities": "2.57.0",
     "@memberjunction/ai": "2.57.0",
+    "@memberjunction/ai-core-plus": "2.57.0",
     "@memberjunction/aiengine": "2.57.0",
     "@memberjunction/ai-prompts": "2.57.0",
     "@memberjunction/doc-utils": "2.57.0"
