+++ packages/MJServer/package.json
@@ -24,6 +24,7 @@
     "@graphql-tools/utils": "^10.0.1",
     "@memberjunction/actions": "2.57.0",
     "@memberjunction/ai": "2.57.0",
+    "@memberjunction/ai-core-plus": "2.57.0",
     "@memberjunction/ai-agents": "2.57.0",
     "@memberjunction/ai-mistral": "2.57.0",
     "@memberjunction/ai-openai": "2.57.0",
